Closes Issue #1 - implement in Golang (WIP)
This is almost a 1:1 port from Python, using, hopefully,
adequately idiomatic Go.

Note that in the Golang implementation we explicitly use concurrency
primitives like mutexes. This is likely to be a source of bugs, the
specific boundary which causes a mutex to be tripped needs to be
assessed with care. The Python prototype does *not* use concurrency,
primarily due to indifference.

Note that the CICD is set up to deliver Golang test reports, rather
than merging other languages' test reports.